Handle missing wayland display gracefully
When the wayland socket didn't exist, the appman-launcher-qml would
terminate with an abort. This could happen for instance inside a
container when the socket file was not mapped. This has been
mitigated to only show a critical error message in the am.runtime.qml
logging category and exit gracefully (with an error return code).

Pick-to: 6.5 5.15
Change-Id: Id29d0e79e260726d1ffd9b940eac806b483b8a19
Reviewed-by: Robert Griebl <robert.griebl@qt.io>