Handle missing wayland display gracefully
When the wayland socket didn't exist, the appman-launcher-qml would
terminate with an abort. This could happen for instance inside a
container when the socket file was not mapped. This has been
mitigated to only show a critical error message in the am.runtime.qml
logging category and exit gracefully (with an error return code).

Change-Id: Id29d0e79e260726d1ffd9b940eac806b483b8a19
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
(cherry picked from commit 61e91284ade4ba7548e84e6550934aff5bcf1ca2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>