Fix race condition when un-installing running apps
It is very unlikely, but there is a race condition between the
installer thread and the main thread regarding the life time of the
Application object that is currently being uninstalled.
We have a clear backtrace from a production system pointing to this
error, but we cannot reproduce this problem in an unit test.

Change-Id: I290582f270455c64e8653813c5e9d47c294f60e1
Pick-to: 5.15 6.2 6.4 6.5
Reviewed-by: Dominik Holland <dominik.holland@qt.io>