Fix race condition when un-installing running apps
It is very unlikely, but there is a race condition between the
installer thread and the main thread regarding the life time of the
Application object that is currently being uninstalled.
We have a clear backtrace from a production system pointing to this
error, but we cannot reproduce this problem in an unit test.

Change-Id: I290582f270455c64e8653813c5e9d47c294f60e1
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit 3d36fe9fbdd18f635de8c3f01b3067d37e3014d7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>