appman-controller: detect broken D-Bus connections and quit with error
In case the AM crashed or the bus died during an install-package
command, the appman-controller would have waited for a taskFail /
taskFinished signal infinitely.
We now detect a D-Bus service owner change (the AM crashed) and a
disconnected D-Bus (the bus died).

Even though the D-Bus daemon dying sounds unlikely, it is the norm
when running the AM with --dbus=auto (the default). If the AM crashes
(or gets terminated by an IDE), the private dbus-daemon just quits.

This makes the controller much more suitable for scripting.

Also fixed a bunch of clang/clazy warnings in this file.

Change-Id: I01a3772fd8773d707984a07d38cbce1d7ab36c94
Pick-to: 5.15 6.5 6.5.0
Reviewed-by: Dominik Holland <dominik.holland@qt.io>