Limit WAYLAND_DISPLAY to control only client side
The WAYLAND_DISPLAY environment variable not only controlled what server
to connect to as a client in a nested setup, but also what server socket
was established. This led to many unusable configurations (either the
socket was already locked or not available). The server socket should be
provided via the CL or configuration file only.

Change-Id: I7fa54c7cc5b5f2baa49b8b68f8de4f98edfc94c0
Reviewed-by: Robert Griebl <robert.griebl@qt.io>