Remove all of the legacy installer user-id separation code
This code is very old and pre-dates modern sandboxing concepts via
kernel namespaces. The only thing it did, was to chown() each
application's installation directory to a unique uid.
The actual runtime did however never setuid() to the app's private
uid when starting it, as this was not easily possible anymore with
the switch to the Sudo-helper forking mechanism.

Neither the configuration, nor the PackageManager properties were
ever documented, so we are effectively removing private API only.

Change-Id: Id4959630fafb408614d1359614ccb74a84e06157
Reviewed-by: Dominik Holland <dominik.holland@qt.io>