qmltestrunner: support BLACKLIST files
Also modernized the TestRunner and made it match declarative's
qmltestrunner a bit more.

The BLACKLIST files live in the test's source directory, so we need to
keep track of that and tell the QTest engine at runtime where the
source directory of the current test is to be found.

QtDeclarative's qmltestrunner has the same problem: the simple
workaround there is to have cmake/ctest run any QMLTEST directly from
the source directory.

Change-Id: I166f8f75a1960d3bb3b899cd645bbd6f3fff41d2
Pick-to: 6.6
Reviewed-by: Dominik Holland <dominik.holland@qt.io>