Split the logging.h header and incorporate the new ColorPrint class
ColorPrint would have originally been introduced with the upcoming
package-server, but this class makes all of the colorized console
logger, the backtrace generator and the StartupTimer output much more
readable, so it was moved to common-lib.

In addition, the per-thread line buffers for the colorized logging
were replaced with a lock-free pool of buffers. There's no point in
assigning a buffer per thread, as we don't have all of them logging
at the same time.

Change-Id: I1aa6664f0e01ebbb14d392fe1d1ede23a2bebfcd
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>