Fix the PackageManager's package role not being usable in strict mode
JS reserves the name "package" in strict mode for future extensions, so
you cannot declare a variable or property with this name, effectively
breaking access to the package role via a required property.

This patch adds a "packageObject" alias role. It also adds similar
roles to ApplicationManager, WindowManager and IntentServer to keep the
model API as similar as possible between the four classes.

Change-Id: Ib9a07cc1107d27c2844367a09a49e6b8c88faf84
Task-number: QTBUG-115998
Pick-to: 6.6
Reviewed-by: Dominik Holland <dominik.holland@qt.io>