Add an ApplicationManagerWindow attached object
This makes it easier to always reach your AMW instance, especially now
in multi-process mode, where Window.window will NOT be the AMW anymore.

The attached object is modelled after the Window attached one, so
activeFocusItem was added to both AMW and AMWAttached.

Also fixed the most obvious of the non-implemented in-process AMW
properties. This needs a follow-up commit to completely fix that
though.

Change-Id: Icac9bfb2994ca6250f0056e142bd59c8ef4b469a
Task-number: QTBUG-103266
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>