DBus cleanup, part 2 and shared QML API
Rewrote the NativeRuntime <-> Launcher communication using auto-
generated interfaces and adaptors.

Also made the shared QML classes (Notification and
ApplicationInterface) simple QObjects, so the QMetaObject is always the
same: both on the AM and the launcher side. This is necessary for qmlsc,
as the compiler doesn't know how to deal with different MOC data for
the same class name in different processes.

Along the same lines, ApplicationManagerWindow is now a QObject, both
for in-process (this was already the case) and out-of process use.
The out-of-process/Wayland implementation will now implicitly allocate
a QQuickWindow and re-parent it accordingly to have the same visibility
behavior as if you would be using normal Windows.

Also FrameTimer got an Impl backend, although in this case a special
implementation is only needed for the System-UI.

And additional unrelated change was rolled into this: launcher-lib was
renamed to application-main-lib and the class LauncherMain to
ApplicationMain to better reflect their purpose (the lib started as the
basis for launcher-qml, but evolved into a generic app/launcher helper)
The reason for including these changes here is that we are touching
all the files in there anyway and having a rename after these big
changes would make git blame far less useful.

Change-Id: Iff915c1b2209f8104051dc131b9add5498df277c
Task-number: QTBUG-103266
Reviewed-by: Dominik Holland <dominik.holland@qt.io>