Make YAML parsing more robust
If libyaml throws an error before it has detected the encoding, we
would instantiate an invalid QTextDecoder to print out the error
context from the YAML. This would then crash due to a nullptr codec.

This patch replaces the codec detection with a hard-coded UTF-8
encoding, the same way the Qt 6 version of the AppMan is doing this.

Yes we loose the ability to parse esoteric UTF16 YAML files, but we
can be sure to have a codec for parsing available.

Change-Id: Ibd4c50bc8162ba8a1d653171d9c55ce024cfff60
Reviewed-by: Dominik Holland <dominik.holland@qt.io>