Fix window mapping test by removing peculiar child window test case
Hiding a child window which parent was made invisible before was not
possible in the multi-process scenario. This behavior was fixed in
qtdeclarative 16023fc77c423a267fcc48894ff942e94cf35b86, so the special
test case for this can be dropped. From now on, the behaviors between
single- and multi-process scenarios are the same and the tests also
reflect that.

Change-Id: Icd17e099d21750ab8971ab54dd6efd9ca7894891
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>