Remove the legacy ApplicationInstaller and clean up "disable-installer"
ApplicationInstaller was deprecated back in 5.13, so it's time to say
goodbye.
This enables us to remove the AM_DISABLE_INSTALLER define from the public API.
Also found and fixed a few locations where the PackageManager
and Sudo singletons were disabled, even though they are independent
from the installer functionality for a long time now.

Change-Id: I1bb0cf2a78500138e998b31641064e4c7f7e26cf
Reviewed-by: Dominik Holland <dominik.holland@qt.io>