Untangle the IntentHandler and IntentServerHandler implementations
Both derive from the same base class, but they are registered into
different QML namespaces, which messes with Creators code model.

Additionally, the docs still said that IntentServerHandler derives
from IntentHandler, which is (a) not true anymore and (b) an
implementation detail.

This commit leaves the base class (it's needed for IntentClient), but
anything related to the QML API is moved into both the derived
classes now.

Change-Id: I6829a5413e88fd932441dd1f84f58c2a98a0edeb
Reviewed-by: Dominik Holland <dominik.holland@qt.io>