Catch first frame in launcher-qml again
This was broken for the most prominent root element, the
ApplicationManagerWindow: both the startup plugin and the timer was
not called any more. For an Item root the startup report was created
twice.

Also removed setting an incubation controller explicitly, since that's
done implicitly, in both appman and appman-launcher-qml.

Got rid of nasty warnings when an ApplicationManagerWindow was closed
in multi-process mode and visibility was set, as well.

Change-Id: I5bad73023dc8e6aaa2262174753a4a5fd5d7d20d
Reviewed-by: Robert Griebl <robert.griebl@qt.io>