Add public defines for QT_AM_VERSION...
We still support building the latest AM against the current Qt LTS, so
we cannot rely on Qt version == AM version.
In order to cope with the latest API changes, we need a way for user
code to check the AM version it compiles against.

Using the same naming and conventions as the QT_VERSION... macros
allows us to do "#if QT_AM_VERSION < QT_VERSION_CHECK(6, 7, 0)"

These are available for custom appman and launcher builds, as well as
for plugins.

Change-Id: If36c3bf80cc8263fad8fcc98d14e3d056fc1fd6d
Reviewed-by: Dominik Holland <dominik.holland@qt.io>