Fix a potential crash, handling an intent request during app removal
Handling/disambiguating an incoming intent request on the server could
lead to accessing a dangling Intent pointer, if the application
serving that Intent was removed during that time period.
The solution is to run the whole handling and disambiguation step on
copies of the Intent pointers. If an app DOES get removed during
handling, the intent call will now fail gracefully, because we either
cannot start the (now missing) application or we cannot establish an
IPC connection anymore.

Change-Id: I2bddf653304f37205c42bd430df448b46f73169a
Pick-to: 6.6 6.5
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>