doc: ApplicationManagerWindow properties and signals
The class documentation still needs a rewrite.

Also in this patch: the inProcess property was renamed to singleProcess
to match the existing property in ApplicationManager. Besides, this
should be a lot clearer to the user, as the documentation only ever
talks about single- vs. multi-process, but never mentions the term
"in-process".

Change-Id: I3baa4e000c8acdb662f2ed472387cdf7e8741048
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit b76b6f0735b0197996ea3c9db111cc3234d20c43)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>