Improve the code-coverage setup
Instead of making a standalone_test build and referencing the
build folder from there, we do the test build together with
the module build and run it like on a developer machine.

This makes the whole report more reliable and fixes some
areas which were marked as not tested at all.

Change-Id: Ifa646ebc613e35bb8aa2dc6e1c1df5c57191c180
Reviewed-by: Robert Griebl <robert.griebl@qt.io>