Fix application quit behavior in single-process mode
Single-process mode QML applications had the finish() callback invoked
twice, if a fade-out animation was running on a window.

While at it, fixed the 'kill' stop flag to be more in line with the
multi-process case and also updated the debug output on application
stop to the newer, more readable style used by the native runtime.

Change-Id: Iee44526d95b1416b213b897fb2b0542d2535c5e7
Fixes: QTBUG-123420
Pick-to: 6.7 6.6
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>