Fix D-Bus marshalling for taskRequestingInstallationAcknowledge
When a new package was installed, the D-Bus marshalling code for the
signal tried to serialize the contained applications' metadata via the
ids, but at that point, the ids are not known to the ApplicationManager
yet.

Instead we now do the same as in the PackageManager class: add a
function that can serialize an Application * (that doesn't need to be
known to the ApplicationManager).

Change-Id: I5989c1e1d449d38d634fd00e7ed33bf66e9430a3
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
(cherry picked from commit 5146513f6bebeabc119d5ec330432b9f4ff61b1a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 7829162b8ca7745abea90089038023c400e43b32)
Reviewed-by: Robert Griebl <robert.griebl@qt.io>