A better watchdog for Wayland, main thread and scene graph
This replaces the existing Wayland ping/pong watchdog with a new setup
that monitors:
 - Wayland XDG clients (on a per-client basis, not per-surface as
   before)
 - The main event loop
 - All scene-graph render threads, split up into the states: syncing,
   rendering and swapping.

Each of those sub-watchdogs has a configurable check interval, a warn
timeout and a kill timeout.

In addition, this watchdog can also handle watchdog request coming from
systemd, if the AM is started via a systemd unit with the relevant
options enabled.

The documentation will come in a follow-up commit.

Fixes: QTBUG-125986
Change-Id: I3d50a0d33625c180965eed98b24c80dc7cf9dece
Reviewed-by: Dominik Holland <dominik.holland@qt.io>