Warn users about wrongly instantiating QML types
Technically we cannot hide SystemUI types in an Application context
and vice versa in single-process mode.
To make matters worse, we also cannot reasonably check each invocation
of a function or property accessor for singletons.

But we CAN at least make sure that the instantiatable types are not
instantiated in the wrong context.

Change-Id: I44ea14f59578f1b3e5230ccfb52e0358896d3d37
Reviewed-by: Dominik Holland <dominik.holland@qt.io>