Fix data race on auto generated DLT id
The "unique" number (which wasn't unique and was only ever used to
generate a DLT id for apps) mechanism had a data race on construction
due to the package scanner working multi-threaded.

Also, the old mechanism tried to be clever and keep the id constant
across restarts, but a simple --clear-cache would wipe the ids anyway.

Projects using DLT will want to set a custom DLT id on apps, so we
just need a simple mechanism to make sure that all apps have a unique
id per session.

Change-Id: I6a7d117a3e941677239babd50af210203170b42a
Pick-to: 6.8
Reviewed-by: Dominik Holland <dominik.holland@qt.io>