Fix data-race on taskRequestingInstallationAcknowledge parameter
The backing PackageInfo* for the 'package' parameter was modified in
the installation thread (setBaseDir), AFTER the signal was emitted
on the UI thread and QML was able to access the baseDir property.

The idea of using a temporary Package object for that signal emission
was not a wise choice, but we cannot easily change that now without
breaking the QML API for PackageManager.

Change-Id: Ib042292e11b42a89263360e501fe7eb90a4ca765
Pick-to: 6.8 6.7 6.5 6.2 5.15
Reviewed-by: Dominik Holland <dominik.holland@qt.io>