Fix data-race on taskRequestingInstallationAcknowledge parameter
The backing PackageInfo* for the 'package' parameter was modified in
the installation thread (setBaseDir), AFTER the signal was emitted
on the UI thread and QML was able to access the baseDir property.

The idea of using a temporary Package object for that signal emission
was not a wise choice, but we cannot easily change that now without
breaking the QML API for PackageManager.

Change-Id: Ib042292e11b42a89263360e501fe7eb90a4ca765
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit a3f04c5ff678b50d93beb8401f226224fd555774)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit f2ca218d9aa37010f357b8a548295a5cbf534269)
Reviewed-by: Robert Griebl <robert.griebl@qt.io>