Watchdog: fix TSAN warnings and make output consistent
We do rely on the Watchdog object being destructed by making it an
child of qApp. The problem here is that we are now in a situation,
where the qApp's vptr has been changed to make it a QCoreApplication
class, while the watchdog thread's event loop is still running and
accessing the same QCoreApplication instance. Although this shouldn't
be a problem in our case, it's still a TSAN warning that can be avoided
by hooking in to aboutToQuit() and tearing down the watchdog thread
there.
(see https://github.com/google/sanitizers/wiki/ThreadSanitizerPopularDataRaces#data-race-on-vptr)

Also, changed the output to reflect that we are not watching threads
directly, but their event loops.

Change-Id: Ie8efd0cc870b28ed6f99187c1cf880903e53b063
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit 4cfe79f0b5e196664503d9ea5a0baa1e4e0154e9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>