Set applicationName and organizationName on apps
All apps started via appman-launcher-qml used to have the same
QCoreApplication applicationName and organizationName. This can lead to
problems when using components that use functionality like
QStandardPaths to create unique per-app data storage. One example of
this is QtInsights.

This patch will now use the applicationId from the apps manifest to set
an applicationName. If the the id is using the recommened reverse-DNS
notation, the organizationName and organizationDomain will also be set
by reversing the reverse-DNS domain part of the applicationId.

A few code blocks in launcher-qml.cpp have been re-arranged to make the
code more readable and not spread the id handling all over the place.

Change-Id: Ibf0cc02aabd8d04b0d6129445128622fca06e597
Pick-to: 6.8
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>