Watchdog: fix the event loop stuck durations being way off
We have been running a timer on the watched thread's event loop to
detect freeze conditions. In addition to QTimers being quite coarse
in general, the reported freeze durations were inherently off by
anything between 0ms and the timer's interval.

The new implementation works the same way as the render loop watchdog:
we register a synchronous callback that wraps every event delivery.
This way we get the exact durations for freezes.

This also opened up the opportunity to simplify the shared data
structures by a lot. Instead of writing to the *Data structures from
the watched threads, we now instead just invokeMethod to the watchdog
thread every time something needs to be reported and then serialize
all writes in that thread without the need for any locking or atomics.
A nice side effect is that the only atomics left are the duration
timers, but now we can just access them as the 64bit values that they
are and we don't have to deal with the 30bit overflow anymore.

As no logging is now done from the watched threads anymore, the split
logging into 2 categories has been simplified and merged.

Change-Id: I0079f2bcb746721f3305d56249962f2bb614b452
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
(cherry picked from commit 759c7c02507a299a8fa9d35801bb4e5cda20414c)
Reviewed-by: Robert Griebl <robert.griebl@qt.io>