Do not compile the build directory into the appman binary
This was a hack to make it possible to always use the freshly compiled
runtimes, when building against a installed Qt, but without installing
the AM itself into the Qt directory.
Because we already put the appman's binary directory in the search path,
even this niche use-case does still work.

Change-Id: Ia53b89d2fa7f77ac96eb17f9493e9dbd2569baaa
Pick-to: 6.8
Fixes: QTBUG-130117
Reviewed-by: Dominik Holland <dominik.holland@qt.io>