Crash handler: make apps report the actual fatal signal
Apps that were killed by a Unix signal did just show up as 'aborted' in
the system-ui, regardless of the actual signal that triggerd the kill.

Also, use the official Unix way to disable core dumps. Any fatal signal
will dump core, if enabled in the system (not just SIGABRT).

Change-Id: Ie5e1ec59eebeaf446a9568d7887a358c2a3cbea0
Pick-to: 6.8
Reviewed-by: Dominik Holland <dominik.holland@qt.io>