Crashhandler: reinstall ourselves, if our handlers got disconnected
One example is Squish which resets ALL signal handlers to default, if
it is instructed to NOT handle signals (this is actually a bug in
Squish).
Since we have to live with these kinds of tools, the better option is
to check at runtime, if the AM's handler got disabled.

Pick-to: 6.8
Change-Id: Idd122adc89130a869b1a37cc6c343f07c285c46c
Reviewed-by: Dominik Holland <dominik.holland@qt.io>