Examples: Fix application-features on Windows
Our cmake setup on Windows creates all DLLs in the top-level build
directory, but we explicitly need the app modules in the respective
app directories.

Also, the Widgets app does not make sense (and cannot work) in
single-process mode, so there's no point in building it.

The Glitches app in addition was not able to import the additional
module, because it was named the same as the app module. Instead of
renaming the second module, the functionality was folded into the
app module in order reduce confusion: the Crash app already shows
how to have another private module inside the app.

With no Wayland available, the Compositor app doesn't make sense, so
the actual implementation got moved into a component and the app now
just shows an error message if that component cannot be loaded.

Change-Id: I3812d4e1205264c73f36a74e573477f70ed3be7b
Pick-to: 6.8 6.8.1
Fixes: QTBUG-130388
Reviewed-by: Dominik Holland <dominik.holland@qt.io>