Update the packaging to be compatible with bsdtar
Both bsdtar and gnu tar write a short block as the last block in a
tar archive.
In addition, fixed the error handling when closing archives: after
archive_write_free() the 'ar' pointer is dead, so we need to split
the code into an explicit _close() and a _free().
Also output the full packager path in the create-test-package.sh
script: this makes it easier to debug CI runs.

Change-Id: I586213387ca253b93a6335059f15af98afde4478
Pick-to: 6.8 6.9
Reviewed-by: Dominik Holland <dominik.holland@qt.io>