Fix potential problems with GNOME platform theme loading
When running the AM on the desktop, the apps "inherit" the current
desktop setting (KDE, GNOME, etc.) and they will initialize the
corresponding platform theme. This is unnecessary and it will also lead
to problems, if that theme cannot initialize itself correctly:
The GNOME theme for example sends out a sync dbus call that can never
be replied to, because AM apps run on a separate session DBus.

Change-Id: I73589b2c20aab4bdecdf561f5b3c0256d8a186ae
Fixes: QTBUG-136234
Pick-to: 6.8
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit 9fff970134d7331ddd1619ceeb815851d686d861)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>