Fix regression in WindowItem.objectFollowsItemSize
Introduced in commit 4e00df28c253f3a92cd4d17a464418261f54c602

This optimization strategy is still valid, but we didn't account for
WindowItems that did not start life with a valid "window" property.
The solution is to keep the "m_completed" flag in the WindowItem
itself, and not in the volatile "m_impl" object.

Change-Id: I20bcbecafc33260a95c116ba5074ed3dd1a7d052
Fixes: QTBUG-137056
Pick-to: 6.10 6.9 6.8
Reviewed-by: Dominik Holland <dominik.holland@qt.io>