packager: Add a new pre-package-command option
The pre-package-command can preprocess each file, before
they are packaged.
This is needed in order to write IMA file signature
and make sure permission and ownership are correct.

The CMake integration also got a new FAKEROOT
option to run the packager with fakeroot, as some
of those modifications need root priviledges.

Change-Id: I883893a839a9c44163d604f1c614b8013c840b8c
Reviewed-by: Robert Griebl <robert.griebl@qt.io>