Do not crash on slow systems during shutdown
Squish will spin a local event loop in aboutToQuit(), which can
lead to recursive call of aboutToQuit(), which then crashes in the
Watchdog code.

We try to mitigate this in two ways:
1) the Watchdog will only react to this signal once
2) the "hang detector" timer in the shutdown machinery will now
   directly call ::exit to avoid any recursion in Qt code

Making the hardcoded timer configurable will be done in a separate
commit, as this cannot be cleanly backported.

Change-Id: Icd16d302510ed73ddc321fa9d877f315f927f854
Pick-to: 6.10 6.9 6.8
Reviewed-by: Dominik Holland <dominik.holland@qt.io>