Tests: workaround for Wayland event loop bug
This auto test has been failing every now and then on Coin, but I
could not reproduce this locally until now:
I can see the xdg_toplevel.close() call being sent by server, but
sometimes the client does not receive it all.

This has to be fixed in QtWayland, but for now we have a workaround
to make this auto test work again.

Change-Id: I8afb1c39cdf99bb3400001ad6e874018899db9a8
Pick-to: 6.10 6.9 6.8
Reviewed-by: Dominik Holland <dominik.holland@qt.io>