Notify the user, if the kernel's AT_SECURE flag is set
This is normally the result of running an executable with the suid bit
set, but it could also be set by a kernel security module.

If set, this can lead to weird, hard to debug behavior, so it's best to
know about this flag being set early on.

Pick-to: 6.9 6.8
Change-Id: I34aa4b405de5ebfcc7bfd04cd8d761dee5e20020
Reviewed-by: Dominik Holland <dominik.holland@qt.io>
(cherry picked from commit 5b9562d1b6b639233f3f5fecc52ed56766588092)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>