Fix quick-launcher not obeying cpu load setting
There were 2 timers runnng in parallel to trigger the quick-launcher
rebuild, and one of them was not checking the current cpu load.

Also fixed the documentation to reflect that an idleLoad of 0 actually
means "no checking".

In addition, the first read of cpu (and also io) load was wrong, as
these are delta values, so we have to implicitly read and eat the
first value read.

Change-Id: I29bb8409a73378b0d908994ccfde467c4204072e
Pick-to: 6.9 6.8 6.5
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
(cherry picked from commit fef3756dacd214940a49334f20b1a858401045b7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>