Security: make sure notifications cannot fake their app id
As the Notifications API is on the session bus, anybody with access
to the bus could create a notification in the name of any other
application (this is a flaw in the original freedesktop protocol).

We can do better though, and correlate the sender's PID with the known
PID of the given applicaton id.

Fake notification requests are now blocked and a critical message is
printed.

Change-Id: I5337bec0ce99165e4151a8b9ff49c0a7f99f266c
Pick-to: 6.9 6.8 6.5
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
(cherry picked from commit 731cd3200da7aa3ee95fa1df1afca59b6845a1b2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>