Do not enable external D-Bus interfaces by default
By moving the notifications to a private interface on the per-app P2P
bus, we can now get rid of the "auto" session-bus that had to be
started to enable notifications.

This "auto" mechanism also registered all the other singleton
interfaces on this session-bus, giving every app access to all the
appman functionality by default.

This patch now gets rid of the "auto" session-bus completely and
defaults the dbus registrations to "none", meaning the user has to
explicitly enable access to the singletons via the configuration.

Also, the p2p development bus is now only started, when the
developmentMode is explicitly set to 'system' or 'application'.

Change-Id: I2335d89c02c9b34b0804d3770ba2afd1c226b774
Reviewed-by: Thomas Senyk <thomas.senyk@qt.io>
Reviewed-by: Dominik Holland <dominik.holland@qt.io>