Cleaner shutdown implementation
The old implementation used static for everything. This made the
tst_main auto-test only work correctly by accident, as it calls
shutdown() multiple times.

Also removed the logging for the shutdown reason when we are running
as a unit test, as this shows up as an "unknown test function" in
the logs.

Change-Id: Ifd21218ed32840756500dd5df595865cc70850b7
Reviewed-by: Dominik Holland <dominik.holland@qt.io>