YAML: Add support for version 1.2 document parsing
YAML 1.2 was specified in 2009 and 1.1 is considered legacy nowadays.

The default for parsing and emitting unversioned documents is still
1.1, but you will see warnings for unversioned documents that use
features that are not compatible with 1.2.

Starting with AppMan 6.12, the default will switch over to 1.2 for
unversioned documents.

In addition all tags from the core schema are now correctly applied
to both 1.1 and 1.2 documents.

Version 1.1 will still be supported in the future for explicitly
versioned documents.

Change-Id: I6c9920784c8f7650d7e63f2273773d7fa462fe9a
Pick-to: 6.11
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>