Fix long-standing race condition in the Unix signal handler
The available signal handlers were stored in a list, but this is
inherently not async-signal-safe. It was giving us the flexibility to
have multiple signal handlers installed for a single signal, but in
practice this is not a requirement at all and was never used.

The new solution is to use a static array of SigHandler*, one handler
per signal (ideally, we would store each std::function inside an
atomic, but that is not possible).
Removal of signal handlers is tricky even with atomics, because signal
handlers block the current thread completely. This means we cannot
delete a removed SigHandler instance immediately, but we can safely
do so after a grace period (currently set to 10min).

Also marked all libc and kernel functions with '::', as is the case
in the rest of the AM source code.

Change-Id: Ibd79fd489dc9d4db8edbf569f7ce29dee30c8d02
Pick-to: 6.11
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>