Change all appman libraries to use Qt's default regarding shared/static
The idea of using static libraries only was to shave off any
milli-second possible from the startup time. This does bring a lot of
problems though, in that a lot of valid use-cases where you need
access to an internal symbol are not possible, because you cannot link
to the static libraries from a Qt or QML plugin, because you get
duplicate symbols.

As we need the possibility to access Appman symbols from auto-generated
plugins for a new intent overhaul, we are doing 2 things now:

1) This patch will simply convert the existing libs over to building
   as static/shared, depending on how Qt itself was built.

2) The follow-up patch will consolidate the library zoo into more
   managable and logical chunks. Plus this will also help with keeping
   the hit to the startup performance in check.

A few destructors used in std::unique_ptrs had to made explicit to make
MSVC happy.

Change-Id: I1bdbb6d8f4c16a6bf70bd98b9a146203d9036a57
Pick-to: 6.11
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>