Re-organize the appman libraries into more manageable chunks
This is part 2 of this patch series. Part 1 was about converting the
libs from static to shared.

This patch now consolidates the library zoo into more manageable and
logical chunks. Plus this will also help with keeping the hit to the
startup performance (incurred by part 1) in check.

The new setup closely follows the QML module nomenclature, but needs
to account for the tool builds as well.

The new libs are:
- Common      no changes
- Package     a combination of Package, Application and Crypto
- Shared      renamed from SharedMain and combined with Monitor and
              IntentClient
- SystemUI    renamed from Main and combined with Manager, DBus,
              IntentServer and Window
- Application renamed from ApplicationMain

Because the CrashHandler is using a "constructor" function, it had to
be moved from the Common to the Shared lib to avoid having the custom
crash handler also in the tools, as they need to install their own
CTRL+C handler at times.

In addition the 3 QML import modules are now built as official QML
plugins and they are installed into $QTDIR/qml as any other import
library.

NB! There is a problem we still have to solve: custom appman binaries
(system-ui's) are currently built by linking against Qt::AppManMain,
but this no longer exists. It would be ideal if we could add an alias
here. The other renamed modules are not that problematic, as they
shouldn't be used outside of the appman project itself.
If an alias is not possible, a dummy library which depends on the new
"SystemUI" one should also do the trick.

Change-Id: Id07e05a523c48e773c295c5be2f27804229155e0
Pick-to: 6.11
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>