Make ProcessTitle work without hacks
We have been using well known hacks to get around the limitation of
Linux not being able to simply change a process' title, like the BSDs
can.
These hacks are prone to fail, with failures happening on arm64,
especially when combined with bwrap containers.

This commit does away with those workarounds and uses an officially
supported mechanism of the Linux kernel to change the memory mapping
for the command line arguments.
In addition, the program_invocation_short_name (used for journal
logging is also changed)

However, this mechanism only works on kernel >= 3.18 and only if
CONFIG_CHECKPOINT_RESTORE is enabled.

Change-Id: I9691a0c2aff767719a20233b262bd5c26758a6f0
Reviewed-by: Dominik Holland <dominik.holland@qt.io>