Refactor packager and controller using a command base class
While at it, the new parsePassword function from the controller was
moved into the base class as well and put to use in the packager as
well.
It also received a secure Windows and Unix implementation for
readPasswordFromConsole(), that do not echo the password while you are
typing.

Change-Id: I163b89c253fae2bcc9e558793487e8f886a37ea3
Reviewed-by: Bernd Weimer <bernd.weimer@qt.io>
(cherry picked from commit ab1f5d04df4a901fe22478cfe433dfc19e7018d6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>