Fix hiding ApplicationManagerWindow
In single-process mode, when an ApplicationManagerWindow's visible
property was set to false, it would still be shown (see mindesk app1
popup for instance).

In addition, in the minidesk and application-features example the
application window containers on the System UI side will follow the
windows' content state now: they will only be shown, when the content
state is SurfaceWithContent. This better reflects the visible state
(basically, the WindowItem doesn't need to and the "chrome" usually
shouldn't be shown, when the WindowObject is invisible).

Pick-to: 6.10
Change-Id: I8ec0f8dd74d71fa287919690881d70a973dd974b
Reviewed-by: Robert Griebl <robert.griebl@qt.io>
(cherry picked from commit 5b0edee37e2f30e1cb74a349f9e899b48a3f1cdf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>