Remove documentUrl from manifest definition
The documentUrl in manifest files was broken:
The documentation for the am-package format only refers to
applications:runtimeParameters:documentUrl, which is made part of the
ApplicationObject.runtimeParameters on System UI side, but otherwise
ignored (in contrast to the documentation not "passed onto, and
interpreted by the application's runtime").
The documentUrl was only evaluated (and used as default when start
application was called), when it was set on the root of a legacy
am-application manifest file. However this contradicts the
ApplicationObject.documentUrl documentation: "Always returns an empty
string".

Pick-to: 6.11
Change-Id: Iac32687456e650b128817eace3551fb182623a82
Reviewed-by: Robert Griebl <robert.griebl@qt.io>